% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_setup.R
\name{module_setup_server}
\alias{module_setup_server}
\title{Server for setup module}
\usage{
module_setup_server(id, module_name)
}
\arguments{
\item{id}{Internally used}

\item{module_name}{Name of the module}
}
\description{
Server for setup module
}
