% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_setup.R
\name{module_setup_server}
\alias{module_setup_server}
\title{Server for setup module}
\usage{
module_setup_server(input, output, session, module_name)
}
\arguments{
\item{input}{Internally used}

\item{output}{Internally used}

\item{session}{Internally used}

\item{module_name}{Name of the module}
}
\description{
Server for setup module
}
